(fix) add int and int64_t casts and minor formatting

+++ hummingbot/core/utils/tracking_nonce.py
@@ -7,6 +7,6 @@ _last_tracking_nonce: int = 0
 # not good enough to create unique order_id.
 def get_tracking_nonce() -> int:
     global _last_tracking_nonce
-    nonce = (time.time() * 1e6)
+    nonce = int(time.time() * 1e6)
     _last_tracking_nonce = nonce if nonce > _last_tracking_nonce else _last_tracking_nonce + 1
     return _last_tracking_nonce
